refactor: make reusable activate/install plugins mechanism of starter sites dashboard
